TTL化のための変更

+++ backend/src/main/java/com/kotonekanno/movielog/controller/AuthController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-import com.kotonekanno.movielog.service.UserService;
+import com.kotonekanno.movielog.service.application.UserService;
 
 import java.util.Map;
 
@@ -19,7 +19,7 @@ public class AuthController {
   }
 
   @PostMapping("/register")
-  public ResponseEntity<Map<String, Long>> register(
+  public ResponseEntity<Map<String, Integer>> register(
       @RequestParam String email,
       @RequestParam String password
   ) {
